refactor: extract contest prefix dictionaries to prefixes.ts
